+++ lakeapi/__init__.py
@@ -4,4 +4,4 @@ __author__ = 'Jan Skoda'
 __email__ = 'skoda@jskoda.cz'
 __version__ = '0.19.1'
 
-from .main import load_data, list_data, available_symbols, set_default_bucket, use_sample_data, cache, set_cache_size_limit # noqa
+from .main import load_data, list_data, available_symbols, set_default_bucket, use_sample_data, used_data, cache, set_cache_size_limit # noqa